+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [0.7.0] - 2025-12-15Z
+
+- [#23](https://github.com/dbdrive/beiwagen/pull/23): Remove legacy output functionality - [@blackjacx](https://github.com/blackjacx).
+- [#22](https://github.com/dbdrive/beiwagen/pull/22): Support Age Ratings - [@blackjacx](https://github.com/blackjacx).
+
 ## [0.6.0] - 2025-10-23Z
 
 - [#21](https://github.com/dbdrive/beiwagen/pull/21): Support request output types - [@blackjacx](https://github.com/blackjacx).